feat(dpp): retours PR
