+++ front/screens/index.ts
@@ -2,7 +2,9 @@ import EpdsFooter from "./epdsSurvey/epdsFooter.component";
 import EpdsGenderEntry from "./epdsSurvey/epdsGenderEntry.component";
 import EpdsQuestion from "./epdsSurvey/epdsQuestion.component";
 import EpdsResult from "./epdsSurvey/epdsResult.component";
-import EpdsResultInformation from "./epdsSurvey/epdsResultInformation.component";
+import EpdsResultContactParagraph from "./epdsSurvey/epdsResultInformation/epdsResultContactParagraph.component";
+import EpdsResultInformation from "./epdsSurvey/epdsResultInformation/epdsResultInformation.component";
+import EpdsResultSimpleParagraph from "./epdsSurvey/epdsResultInformation/epdsResultSimpleParagraph.component";
 import EpdsSurveyContent from "./epdsSurvey/epdsSurveyContent.component";
 import EpdsSurveyScreen from "./epdsSurvey/epdsSurveyScreen.component";
 
@@ -11,7 +13,9 @@ export {
   EpdsGenderEntry,
   EpdsQuestion,
   EpdsResult,
+  EpdsResultContactParagraph,
   EpdsResultInformation,
+  EpdsResultSimpleParagraph,
   EpdsSurveyContent,
   EpdsSurveyScreen,
 };
